feat(finance): extract payments list into feature components (Phase 6e)
- Extract Payments list logic into PaymentsTable feature component
- Extract table JSX into dedicated PaymentRow component
- Enforce keyboard accessibility (role="button", onKeyDown) for payment rows
