+++ features/payments/index.ts
@@ -1,0 +1,2 @@
+export * from './components/payments-table';
+export * from './components/payment-row';
